requirements.txt
langchain==0.0.154
openai==0.27.5
streamlit==1.22.0
black==23.3.0
tabulate==0.9.0
python-dotenv